Switch square and curly brackets
This breaks the (artificial) symmetry of the curly braces being "above"
the square ones, as they are on QWERTY, and their relation to the `\,|`
key.

But the originally shifted curly brackets being on the home row
while the square brackets required moving the fingers to the lower row
didn't feel super intuitive.